consolidated test classes
